Remove horizontal space between bar graph and container
Fix percentage calculation in bar graphs

Fixed layout of demographics.

Disable touch events on clusters
