+++ src/App.scss
@@ -420,16 +420,42 @@ h6 {
     }
   }
 
+  .meta-secondary {
+    display: flex;
+    flex-direction: row;
+    align-self: center;
+    margin-top: 3rem;
+    margin-bottom: 3rem;
+    .unknown {
+      display: flex;
+      flex-direction: row;
+      width: 6rem;
+      font-family: 'archia';
+      font-size: 10px;
+      background: $gray-light;
+      padding: 0.25rem;
+      border-radius: 5px;
+      color: $gray;
+      svg {
+        width: 36px;
+        margin-right: 0.25rem;
+        stroke-width: 2px;
+        color: $gray-mid;
+        margin-top: -0.25rem;
+      }
+    }
+  }
+
   .district-bar, .to-essentials {
     width: 100%;
   }
 
-  .Minigraph, .Level, .MapExplorer {
+  .Minigraph, .Level, .MapExplorer, .meta-secondary {
     width: calc(100% - 3rem);
   }
 
   .MapExplorer {
-    .header, .map-stats, .back-button, .last-update {
+    .header, .map-stats, .back-button, .last-update, .unknown {
       display: none;
     }
   }
@@ -1059,6 +1085,7 @@ h6 {
   .suggestions {
     display: flex;
     flex-direction: column;
+    margin-right: 1rem;
     .suggestion {
       display: flex;
       flex-direction: row;
@@ -3370,6 +3397,7 @@ select {
     display: flex;
     flex-direction: column;
     width: 10rem;
+    margin-right: 1rem;
   }
 
   .districts {
@@ -3396,6 +3424,7 @@ select {
       h5 {
         margin-top: 0.15rem;
         margin-left: 0.5rem;
+        display: table; 
       }
 
       & > * {
@@ -3431,6 +3460,10 @@ select {
     margin-bottom: 0;
     width: 100%;
     padding-left: 1.5rem;
+  }
+
+  canvas {
+    touch-action: auto !important;
   }
 }
 
@@ -3493,10 +3526,10 @@ select {
     flex-direction: column;
     position: fixed;
     min-width: 5rem;
-    background: $gray-light;
+    background: $gray-light-opaque;
     height: 100%;
     justify-content: flex-start;
-    z-index: 9999;
+    z-index: 99999;
     .navbar-left {
       order: 3;
       margin-top: auto;
@@ -3543,6 +3576,10 @@ select {
         font-weight: 600;
       }
     }
+  }
+
+  .PatientsDB {
+    margin-left: 5rem;
   }
 }
