+++ src/components/state.js
@@ -130,6 +130,19 @@ function State(props) {
               }
             </React.Fragment>
           )}
+
+          {fetched && (
+            <div className="meta-secondary">
+              <div className="unknown">
+                <Icon.AlertCircle />
+                <div className="unknown-right">
+                  Awaiting district details for{' '}
+                  {districtData[stateName].districtData['Unknown']?.confirmed}{' '}
+                  cases
+                </div>
+              </div>
+            </div>
+          )}
         </div>
 
         <div className="state-right">
@@ -177,8 +190,7 @@ function State(props) {
                   {
                     <DeltaBarGraph
                       timeseries={timeseries.slice(-5)}
-                      key1={'dailyconfirmed'}
-                      key2={'totalconfirmed'}
+                      arrayKey={'dailyconfirmed'}
                     />
                   }
                 </div>
